requirements.txt
python
asyncio
logging
MetaTrader5
telethon
python-telegram-bot
python-dotenv
aiohttp
aiofiles